refactor(tests): пофайловые фейки досок и пороги в ProviderAdapter
