+++ tests/tasks/boards/fakes/__init__.py
@@ -1,0 +1,1 @@
+"""Пофайловые фейки досок для общего contract-набора."""
